backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic-settings==2.7.0
pillow==11.0.0
pypdf==5.1.0
reportlab==4.2.5
aiofiles==24.1.0
numpy==2.2.1
rembg==2.0.61
onnxruntime==1.22.1
email-validator==2.2.0
# Optional AI extras for higher-quality upscaling and CUDA acceleration.
# Install if you want the real Real-ESRGAN path enabled:
# torch
# torchvision
# basicsr
# realesrgan
